fix: UDP voice robustness — liveness fallback, crypt gate, resync, stats
Voice was committed to UDP on a purely local socket connect: on a network
where UDP is filtered it went into a black hole with no recovery, socket
errors tore down the healthy TCP session, and a nonce desync silently
killed UDP receive forever. Server-side stats for Brmble clients read zero.

- UDP liveness: voice only goes over UDP while an encrypted ping echo or
  incoming voice packet round-tripped within the last 12s; otherwise it
  uses the TCP tunnel. Pings keep probing, so UDP recovers automatically.
- Crypt gate: SendVoice checks CryptState.Initialized, making the tunnel
  fallback reachable instead of NRE-ing in Encrypt pre-CryptSetup.
- SocketException in UdpSocket send/receive now marks UDP unusable and
  falls back to the tunnel instead of killing the session (ICMP port
  unreachable surfaces as SocketException on Windows).
- Client-initiated crypt resync: after 10 consecutive decrypt failures,
  send an empty CryptSetup (rate-limited to one per 5s) so the server
  re-syncs its nonce.
- TCP Ping now reports UDP stats: crypt good/late/lost/resync counters and
  measured UDP ping RTT avg/var, so the server user-info dialog shows real
  numbers for Brmble clients.
- UDP ping timestamp is now encoded as a proper Mumble varint (0xF4 8-byte
  form) instead of raw fixed 8 bytes.
- Fixed UdpPacketReader 4/8-byte varint decode: shifts were done on int,
  so shift counts were masked (<<56 became <<24) and values >= 2^31
  sign-extended negative.
- Cleanup: removed unused ForceTcp, dead UdpSocket.SendPing, and the empty
  PackVoicePacket stub.

Fixes #595

+++ lib/MumbleSharp/MumbleSharp/UdpPacketReader.cs
@@ -61,12 +61,16 @@ namespace MumbleSharp
                     if ((b & 4) == 4)
                     {
                         //111101__ + long (8 bytes)
-                        return ReadByte() << 56 | ReadByte() << 48 | ReadByte() << 40 | ReadByte() << 32 | ReadByte() << 24 | ReadByte() << 16 | ReadByte() << 8 | ReadByte();
+                        //Shifts must be on long: an int shift count is masked to 0-31,
+                        //so `byte << 56` silently became `<< 24` and scrambled the value.
+                        return (long)ReadByte() << 56 | (long)ReadByte() << 48 | (long)ReadByte() << 40 | (long)ReadByte() << 32
+                             | (long)ReadByte() << 24 | (long)ReadByte() << 16 | (long)ReadByte() << 8 | ReadByte();
                     }
                     else
                     {
                         //111100__ + int (4 bytes)
-                        return ReadByte() << 24 | ReadByte() << 16 | ReadByte() << 8 | ReadByte();
+                        //Widen before shifting so values ≥ 2^31 don't sign-extend negative.
+                        return (long)ReadByte() << 24 | (long)ReadByte() << 16 | (long)ReadByte() << 8 | ReadByte();
                     }
                 case 5:
                     //111110 + varint (negative)
